Move helpers to core
